% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_auth_params}
\alias{add_auth_params}
\title{Add credentials to the API request}
\usage{
add_auth_params(params, api_key = NULL, api_secret = NULL)
}
\arguments{
\item{params}{named list of parameters}

\item{api_key}{client API key}

\item{api_secret}{client API secret}
}
\description{
Add key and secret to the list of parameters for request
}
