% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{build_endpoint_url}
\alias{build_endpoint_url}
\title{Build the API endpoint}
\usage{
build_endpoint_url(endpoint)
}
\arguments{
\item{endpoint}{an API endpoint without leading and trailing slashes}
}
\description{
Take an endpoint and attach it to the base url
}
