% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_call_details.R
\name{rb_call_details}
\alias{rb_call_details}
\title{Call details}
\usage{
rb_call_details(general_call_id)
}
\arguments{
\item{general_call_id}{unique call identifier}
}
\description{
Call details by ID
}
