% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_calls_by_internal_number.R
\name{rb_calls_by_internal_number}
\alias{rb_calls_by_internal_number}
\title{List of recent calls by internal number}
\usage{
rb_calls_by_internal_number(internal_number)
}
\arguments{
\item{internal_number}{employee's internal number}
}
\description{
List of recent incoming and outgoing calls by internal number (up to 50 calls, last 2 weeks)
}
