% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_history_by_customer_id.R
\name{rb_history_by_customer_id}
\alias{rb_history_by_customer_id}
\title{List of calls by Customer ID}
\usage{
rb_history_by_customer_id(customer_id)
}
\arguments{
\item{customer_id}{Customer ID}
}
\description{
List of incoming and outgoing calls by Customer ID
}
