% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_history_by_external_number.R
\name{rb_history_by_external_number}
\alias{rb_history_by_external_number}
\title{List of calls by external number}
\usage{
rb_history_by_external_number(external_numbers)
}
\arguments{
\item{external_numbers}{single external number or vector of numbers}
}
\description{
List of incoming and outgoing calls by external number
}
