% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{send_api_request}
\alias{send_api_request}
\title{Send request to Binotel API}
\usage{
send_api_request(endpoint, params = list())
}
\arguments{
\item{endpoint}{API endpoint specific to the request}

\item{params}{list of parameters for request}
}
\description{
Send POST request with parameters and key/secret to /code{endpoint}
}
